fix: list role note for nav landmark
Just updating based on something I saw from Sara on Twitter and confirmed with some testing on a recent project. Happy to try and find a better link if a Twitter thread isn't the best place to send people. Tried to add a disclaimer to make it clear as realise not everybody might want to be taken there!